Apply db clear fix; customize further

+++ FaceRecognition/MainForm.cs
@@ -93,18 +93,28 @@ namespace FaceRecognition
 
         void ClearTrainedFaces()
         {
+            
+            if (
+                MessageBox.Show( "Essa é uma ação destrutiva.\n\nVocê tem certeza que deseja prosseguir?", "Confirmação", MessageBoxButtons.YesNo )
+                    != DialogResult.Yes
+            )
+            {
+                MessageBox.Show("Ação cancelada com êxito.");
+                return;
+            }
 
             IDataStoreAccess dataStore = new DataStoreAccess(_databasePath);
 
-            var username = Guid.NewGuid().ToString();
-
-            var filePath = Application.StartupPath + String.Format("/{0}.bmp", username);
+            dataStore.ClearTrainedFaces();
 
             foreach (var file in new DirectoryInfo(Application.StartupPath).GetFiles("*.bmp")) {
                 file.Delete();
             }
 
-            MessageBox.Show(dataStore.ClearTrainedFaces(), "Salvar resultado", MessageBoxButtons.OK);
+            img.BackgroundImage = null;
+
+            msg.Text = "Dados foram esvaziados da base.";
+            msg.ForeColor = System.Drawing.Color.Black;
 
         }
 
@@ -119,7 +129,7 @@ namespace FaceRecognition
             try
             {
                 var img1 = WinformUtilities.RecognizeImage(img);
-                imgCamUser.Image = img1;
+                this.img.Image = img1;
             }
             catch (Exception ex)
             {
@@ -137,8 +147,8 @@ namespace FaceRecognition
             var img = WinformUtilities.OpenImageFile();
             try
             {
-                var img1 = WinformUtilities.RecognizeImage( img );
-                imgCamUser.Image = img1;
+                var img1 = WinformUtilities.RecognizeImage( img, msg );
+                this.img.Image = img1;
             }
             catch ( Exception ex )
             {
@@ -155,12 +165,32 @@ namespace FaceRecognition
             if ( imgs == null )
                 return;
 
+            var unidentified = 0;
+
+            var multiple = 0;
+
             foreach ( var img in imgs )
-                processed.Add( WinformUtilities.TrainImage( img ) );
+            {
+                try
+                {
+                    processed.Add( WinformUtilities.TrainImage( img ) );
+                }
+                catch (Exception ex)
+                {
+                    if ( ex.Message == "no_faces" )
+                        unidentified++;
+                    else if ( ex.Message == "multiple_faces" )
+                        multiple++;
+                    else
+                        MessageBox.Show(ex.Message);
+                }
+            }
+
             string regNames = String.Join( "\n", processed.Where( v => v.Length > 0 ) );
+
             MessageBox.Show(
                 regNames.Length > 0 ?
-                string.Format( "Imagens processadas:\n\n{0}", regNames )
+                string.Format( "Imagens processadas:\n\n{0}\n\nImagens sem rosto identificado: {1}\nImagens com múltiplos rostos: {2}", regNames, unidentified, multiple)
                 : "Nenhuma imagem válida processada."
             );
         }
